+++ backend/src/app.ts
@@ -1,9 +1,9 @@
-import pathRoutes from './routes/pathRoutes';
-import userRoutes from './routes/userRoutes';
+import pathRoutes from './routes/api';
+import userRoutes from './routes/api';
 import testRoutes from './routes/testRoutes';
-import filesRoutes from './routes/filesRoutes';
-import hostMsgRoutes from './routes/hostMsgRoutes';
-import compressRoutes from './routes/compressRoutes';
+import filesRoutes from './routes/api';
+import hostMsgRoutes from './routes/api';
+import compressRoutes from './routes/api';
 import express from 'express';
 import cors from 'cors';
 import './config/database'; // 執行`自執行函數`
