+++ backend/src/types/index.ts
@@ -1,0 +1,16 @@
+export interface User {
+  id: number;
+  name: string;
+}
+
+export interface Path {
+  path: string;
+  name: string;
+}
+
+export interface Files {
+  files: string[];
+  zipPath: string;
+  zipName: string;
+  delList: string[];
+}
